updates
1. Currently, some might have to scroll down to see the buy and sell buttons when playing the trading game. Is there a way to ensure all the stuff appears on the same screen?  We would like for the “buy” and “sell” buttons to appear on the same screen as the stocks’ price paths, if possible.

2. Survey questions 1-4 and 15: the blank box only appears after the respondent clicks the radio button. Can we make the blank box initially be shown on the page and disappear only if the respondent clicks the other button?

The "-" in survey questions 2 and 3 should be flipped. Question 2 should have a "-" so that it says "largest loss was -$", and there should NOT be a "-" in question 3 (it currently reads, "largest gain was -$".

At the top of the page with survey questions 5-7, we say "For question 5 to 7". This should say "For questions 5 to 7".

Currently, it’s possible to leave survey questions blank and receive a payment code. Can we force people to answer all the survey questions before they can proceed to the next page?

When clicking “next,” the page view stayed at the bottom of the screen. Can it restart to the top of the screen?